extrac table config to Generic.js
