+++ packages/data-tables/src/components/Table.js
@@ -1,15 +1,4 @@
-import React, { useMemo, useState, useCallback } from 'react';
-import {
-  useFlexLayout,
-  useFilters,
-  useGlobalFilter,
-  useResizeColumns,
-  useSortBy,
-  useExpanded,
-  usePagination,
-  useTable,
-} from 'react-table';
-import matchSorter from 'match-sorter';
+import React, { useState, useCallback } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
@@ -112,101 +101,29 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const Table = ({ columns, data, id, dataForTsv, order }) => {
+const Table = ({
+  columns,
+  data,
+  id,
+  dataForTsv,
+  order,
+  getTableProps,
+  getTableBodyProps,
+  prepareRow,
+  headerGroups,
+  page,
+  rows,
+  canPreviousPage,
+  canNextPage,
+  pageCount,
+  gotoPage,
+  nextPage,
+  previousPage,
+  setPageSize,
+  setGlobalFilter,
+  state: { pageIndex, pageSize, globalFilter },
+}) => {
   const classes = useStyles();
-
-  const filterTypes = useMemo(() => {
-    const storeValuesOfNestedObj = (obj, keyArr) => {
-      for (const key in obj) {
-        if (typeof obj[key] === 'object' && obj[key] !== null) {
-          if (
-            obj[key].class &&
-            obj[key].id &&
-            obj[key].label &&
-            obj[key].taxonomy
-          ) {
-            keyArr.push(obj[key].label);
-          }
-          if (Array.isArray(obj[key]) && typeof obj[key][0] === 'object') {
-            if (
-              obj[key][0].class &&
-              obj[key][0].id &&
-              obj[key][0].label &&
-              obj[key][0].taxonomy
-            ) {
-              keyArr.push(obj[key].map((o) => o.label));
-            } else if (obj[key][0].pato_evidence) {
-              keyArr.push(
-                ...obj[key].map((o) => [
-                  o.pato_evidence.entity_term.label,
-                  o.pato_evidence.entity_type,
-                  o.pato_evidence.pato_term,
-                ])
-              );
-            } else {
-              console.error(
-                'Data is surely array of Object. But it is not Tagtype data.'
-              );
-              console.error(key);
-              console.error(obj[key]);
-            }
-          } else {
-            storeValuesOfNestedObj(obj[key], keyArr);
-          }
-        } else {
-          keyArr.push(obj[key]);
-        }
-      }
-    };
-
-    return {
-      defaultFilter: (rows, id, filterValue) => {
-        const keyFunc = (row) => {
-          let keyArr = [];
-          const rowVals = row.values;
-
-          id.forEach((i) => {
-            if (typeof rowVals[i] === 'object') {
-              storeValuesOfNestedObj(rowVals[i], keyArr);
-            } else {
-              keyArr.push(rowVals[i]);
-            }
-          });
-
-          return keyArr;
-        };
-
-        return matchSorter(rows, filterValue, {
-          keys: [(row) => keyFunc(row)],
-          threshold: matchSorter.rankings.CONTAINS,
-        });
-      },
-    };
-  }, []);
-
-  const defaultColumnFilter = ({ column: { filterValue, setFilter } }) => {
-    return (
-      <input
-        value={filterValue || ''}
-        onChange={(e) => {
-          setFilter(e.target.value || undefined);
-        }}
-        placeholder={`Search...`}
-        type="search"
-      />
-    );
-  };
-
-  const defaultColumn = useMemo(
-    () => ({
-      filter: 'defaultFilter',
-      Filter: defaultColumnFilter,
-      minWidth: 120,
-      width: 180,
-      maxWidth: 600,
-    }),
-    []
-  );
 
   const [isCellExpanded, setCellExpanded] = useState(false);
   const handleCellExpandedToggle = useCallback(
@@ -214,45 +131,6 @@ const Table = ({ columns, data, id, dataForTsv, order }) => {
       setCellExpanded(!isCellExpanded);
     },
     [isCellExpanded, setCellExpanded]
-  );
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    prepareRow,
-    headerGroups,
-    page,
-    rows,
-    canPreviousPage,
-    canNextPage,
-    pageCount,
-    gotoPage,
-    nextPage,
-    previousPage,
-    setPageSize,
-    setGlobalFilter,
-    state: { pageIndex, pageSize, globalFilter },
-  } = useTable(
-    {
-      columns,
-      data,
-      disableSortRemove: true,
-      filterTypes,
-      defaultColumn,
-      globalFilter: 'defaultFilter',
-      initialState: {
-        pageIndex: 0,
-        pageSize: 10,
-        sortBy: [{ id: columns[0].accessor, desc: false }],
-      },
-    },
-    useFlexLayout,
-    useFilters,
-    useGlobalFilter,
-    useResizeColumns,
-    useSortBy,
-    useExpanded,
-    usePagination
   );
 
   const renderIcon = (column) => {
