+++ packages/data-tables/src/components/TableHasGroupedRow.js
@@ -1,16 +1,4 @@
-import React, { useMemo, useState, useCallback, useEffect } from 'react';
-import {
-  useFlexLayout,
-  useFilters,
-  useGlobalFilter,
-  useGroupBy,
-  useResizeColumns,
-  useSortBy,
-  useExpanded,
-  usePagination,
-  useTable,
-} from 'react-table';
-import matchSorter from 'match-sorter';
+import React, { useState, useCallback, useEffect } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
 import ArrowUpwardIcon from '@material-ui/icons/ArrowUpward';
@@ -113,87 +101,34 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const TableHasGroupedRow = ({ columns, data, id, dataForTsv, order }) => {
+const TableHasGroupedRow = ({
+  columns,
+  data,
+  id,
+  dataForTsv,
+  order,
+  getTableProps,
+  getTableBodyProps,
+  prepareRow,
+  headerGroups,
+  page,
+  rows,
+  canPreviousPage,
+  canNextPage,
+  pageOptions,
+  pageCount,
+  gotoPage,
+  nextPage,
+  previousPage,
+  setPageSize,
+  setGlobalFilter,
+
+  // useExpanded
+  toggleAllRowsExpanded,
+
+  state: { pageIndex, pageSize, globalFilter },
+}) => {
   const classes = useStyles();
-
-  const filterTypes = useMemo(() => {
-    const storeValueOfNestedObj = (obj, keyArr) => {
-      for (const key in obj) {
-        if (typeof obj[key] === 'object' && obj[key] !== null) {
-          if (
-            obj[key].class &&
-            obj[key].id &&
-            obj[key].label &&
-            obj[key].taxonomy
-          ) {
-            keyArr.push(obj[key].label);
-          }
-          if (Array.isArray(obj[key]) && typeof obj[key][0] === 'object') {
-            if (
-              obj[key][0].class &&
-              obj[key][0].id &&
-              obj[key][0].label &&
-              obj[key][0].taxonomy
-            ) {
-              keyArr.push(obj[key].map((o) => o.label));
-            } else if (obj[key][0].pato_evidence) {
-              keyArr.push(
-                ...obj[key].map((o) => [
-                  o.pato_evidence.entity_term.label,
-                  o.pato_evidence.entity_type,
-                  o.pato_evidence.pato_term,
-                ])
-              );
-            } else {
-              console.error(
-                'Data is surely array of Object. But it is not Tagtype data.'
-              );
-              console.error(key);
-              console.error(obj[key]);
-            }
-          } else {
-            storeValueOfNestedObj(obj[key], keyArr);
-          }
-        } else {
-          keyArr.push(obj[key]);
-        }
-      }
-    };
-
-    return {
-      defaultFilter: (rows, id, filterValue) => {
-        const keyFunc = (row) => {
-          let keyArr = [];
-          const rowVals = row.values;
-
-          id.forEach((i) => {
-            if (typeof rowVals[i] === 'object') {
-              storeValueOfNestedObj(rowVals[i], keyArr);
-            } else {
-              keyArr.push(rowVals[i]);
-            }
-          });
-
-          return keyArr;
-        };
-
-        return matchSorter(rows, filterValue, {
-          keys: [(row) => keyFunc(row)],
-          threshold: matchSorter.rankings.CONTAINS,
-        });
-      },
-    };
-  }, []);
-
-  const defaultColumn = useMemo(
-    () => ({
-      filter: 'defaultFilter',
-      minWidth: 120,
-      width: 180,
-      maxWidth: 600,
-    }),
-    []
-  );
 
   const [isCellExpanded, setCellExpanded] = useState(false);
   const handleCellExpandedToggle = useCallback(
@@ -201,56 +136,6 @@ const TableHasGroupedRow = ({ columns, data, id, dataForTsv, order }) => {
       setCellExpanded(!isCellExpanded);
     },
     [isCellExpanded, setCellExpanded]
-  );
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    prepareRow,
-    headerGroups,
-    page,
-    rows,
-    canPreviousPage,
-    canNextPage,
-    pageOptions,
-    pageCount,
-    gotoPage,
-    nextPage,
-    previousPage,
-    setPageSize,
-    setGlobalFilter,
-
-    // useExpanded
-    toggleAllRowsExpanded,
-
-    state: { pageIndex, pageSize, globalFilter },
-  } = useTable(
-    {
-      columns,
-      data,
-      disableSortRemove: true,
-      filterTypes,
-      defaultColumn,
-      globalFilter: 'defaultFilter',
-      // initialState: { pageIndex: 0 },
-      paginateExpandedRows: false,
-      initialState: {
-        pageIndex: 0,
-        pageSize: 10,
-        sortBy: [{ id: `${columns[0].accessor}.label`, desc: false }],
-        groupBy: [`${columns[0].accessor}.label`],
-        hiddenColumns: [`${columns[0].accessor}`],
-        expanded: isCellExpanded,
-      },
-    },
-    useFlexLayout,
-    useFilters,
-    useGlobalFilter,
-    useResizeColumns,
-    useGroupBy,
-    useSortBy,
-    useExpanded,
-    usePagination
   );
 
   useEffect(() => {
